requirements.txt
alembic==1.7.3
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
Flask-DotEnv==0.1.2
Flask-JWT-Extended==4.3.0
Flask-Migrate==3.1.0
Flask-Seeder==1.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
postgres==4.0
psycopg2-binary==2.9.1
psycopg2-pool==1.1
PyJWT==2.1.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.25
Werkzeug==2.0.1
yapf==0.31.0

flasgger~=0.9.5
marshmallow~=3.13.0